backend/requirements.txt
Flask
Flask-Cors
pymongo
bcrypt
pyjwt
fastapi
uvicorn
pillow
google-generativeai
python-dotenv
